backend/requirements.txt
# ============================================================================
# CompareIntel Backend - Production Dependencies
# ============================================================================
# This file contains only runtime dependencies needed to run the application.
# For development tools (testing, linting, etc.), see requirements-dev.txt
# ============================================================================

# ============================================================================
# Core FastAPI Framework
# ============================================================================
fastapi>=0.100.0
starlette>=0.27.0
pydantic>=2.0.0
pydantic_core>=2.0.0
pydantic-settings>=2.0.0  # Settings management for Pydantic v2
annotated-types>=0.5.0
typing_extensions>=4.0.0

# ============================================================================
# ASGI Server & Production
# ============================================================================
uvicorn>=0.20.0
gunicorn
click>=8.0.0

# ============================================================================
# Performance Enhancements
# ============================================================================
uvloop>=0.17.0        # Faster event loop (Unix only)
httptools>=0.5.0      # Faster HTTP parsing
h11>=0.14.0           # HTTP/1.1 protocol implementation
websockets>=10.0      # WebSocket support

# ============================================================================
# Request Handling & Utilities
# ============================================================================
python-multipart>=0.0.5  # Form data parsing
python-dotenv>=1.0.0     # Environment variables
PyYAML>=6.0              # YAML configuration
anyio>=3.6.0,<4.0.0      # Async compatibility layer
sniffio>=1.3.0           # Async library detection
idna>=3.0.0              # Internationalized domain names
httpx>=0.27.0            # Async HTTP client (for reCAPTCHA verification)

# ============================================================================
# API Clients & Tokenizers
# ============================================================================
openai>=1.0.0         # OpenRouter API client
tiktoken>=0.5.0       # Token counting for context management
anthropic>=0.18.0     # Anthropic Claude tokenizer (free, no API key needed)
transformers>=4.35.0  # Hugging Face tokenizers for Meta, Mistral, DeepSeek, Qwen
sentencepiece>=0.1.99 # Required by some tokenizers (e.g., Llama, Qwen)
requests>=2.25.0      # HTTP client for utility scripts

# ============================================================================
# Authentication & Security
# ============================================================================
python-jose[cryptography]>=3.3.0  # JWT token handling
passlib[bcrypt]>=1.7.4            # Password hashing (legacy support)
bcrypt==4.0.1                      # Password hashing (primary)

# ============================================================================
# Database
# ============================================================================
sqlalchemy>=2.0.0        # ORM
alembic>=1.12.0          # Database migrations
psycopg2-binary>=2.9.9   # PostgreSQL driver

# ============================================================================
# Email
# ============================================================================
fastapi-mail>=1.4.0      # Email sending
pydantic[email]>=2.0.0   # Email validation

# ============================================================================
# Timezone Support
# ============================================================================
pytz>=2024.1             # Timezone handling for credit resets
